refactoring some of the field code
