+++ nexus_constructor/component_fields.py
@@ -372,3 +372,61 @@ def add_fields_to_component(component: Component, fields_widget: QListWidget):
                 additional_info=str(error),
                 parent=fields_widget.parent().parent(),
             )
+
+
+def update_existing_link_field(field: h5py.SoftLink, new_ui_field: FieldWidget):
+    """
+    Fill in a UI link field for an existing link in the component
+    :param field: The link field in the component group
+    :param new_ui_field: The new UI field to fill in with existing data
+    """
+    new_ui_field.field_type = FieldType.link.value
+    new_ui_field.value = field.parent.get(field.name, getlink=True).path
+
+
+def update_existing_array_field(field: h5py.Dataset, new_ui_field: FieldWidget):
+    """
+    Fill in a UI array field for an existing array field in the component group
+    :param value: The array dataset's value to copy to the UI fields list model
+    :param new_ui_field: The new UI field to fill in with existing data
+    """
+    new_ui_field.dtype = field.dtype
+    new_ui_field.field_type = FieldType.array_dataset.value
+    new_ui_field.value = field[()]
+
+
+def update_existing_scalar_field(field: h5py.Dataset, new_ui_field: FieldWidget):
+    """
+    Fill in a UI scalar field for an existing scalar field in the component group
+    :param field: The dataset to copy into the value line edit
+    :param new_ui_field: The new UI field to fill in with existing data
+    """
+    dtype = field.dtype
+    if "S" in str(dtype):
+        dtype = h5py.special_dtype(vlen=str)
+        new_ui_field.value = field[()]
+    else:
+        new_ui_field.value = field[()]
+    new_ui_field.dtype = dtype
+    new_ui_field.field_type = FieldType.scalar_dataset.value
+
+
+def update_existing_stream_field(field: h5py.Dataset, new_ui_field: FieldWidget):
+    """
+    Fill in a UI stream field for an existing stream field in the component group
+    :param field: The dataset to copy into the value line edit
+    :param new_ui_field: The new UI field to fill in with existing data
+    """
+    new_ui_field.field_type = FieldType.kafka_stream.value
+    new_ui_field.streams_widget.update_existing_stream_info(field)
+
+
+def handle_fields(component):
+    scalar_fields, array_fields, stream_fields, link_fields = component.get_fields()
+    update_methods = [
+        (scalar_fields, update_existing_scalar_field),
+        (array_fields, update_existing_array_field),
+        (stream_fields, update_existing_stream_field),
+        (link_fields, update_existing_link_field),
+    ]
+    return update_methods
